Move DisconnectableInputStream and BulkReadInputStream
