+++ subprojects/core/src/main/java/org/gradle/execution/DefaultCancellableOperationManager.java
@@ -19,7 +19,7 @@ package org.gradle.execution;
 import org.gradle.api.Action;
 import org.gradle.initialization.BuildCancellationToken;
 import org.gradle.internal.UncheckedException;
-import org.gradle.util.DisconnectableInputStream;
+import org.gradle.util.internal.DisconnectableInputStream;
 
 import java.io.IOException;
 import java.util.concurrent.ExecutorService;
