+++ subprojects/core/src/main/java/org/gradle/process/internal/streams/ForwardStdinStreamsHandler.java
@@ -19,7 +19,7 @@ package org.gradle.process.internal.streams;
 import org.gradle.api.UncheckedIOException;
 import org.gradle.internal.UncheckedException;
 import org.gradle.process.internal.StreamsHandler;
-import org.gradle.util.DisconnectableInputStream;
+import org.gradle.util.internal.DisconnectableInputStream;
 
 import java.io.IOException;
 import java.io.InputStream;
