+++ subprojects/core/src/test/groovy/org/gradle/util/internal/DisconnectableInputStreamTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.util
+
+package org.gradle.util.internal
 
 import org.gradle.api.Action
 import org.gradle.internal.concurrent.ExecutorFactory
-import org.gradle.util.internal.MultithreadedTestRule
 import org.junit.Test
 
 import java.util.concurrent.CopyOnWriteArrayList
