+++ subprojects/launcher/src/main/java/org/gradle/launcher/daemon/client/InputForwarder.java
@@ -21,7 +21,7 @@ import org.gradle.internal.concurrent.ExecutorFactory;
 import org.gradle.internal.concurrent.Stoppable;
 import org.gradle.internal.concurrent.ManagedExecutor;
 import org.gradle.internal.io.TextStream;
-import org.gradle.util.DisconnectableInputStream;
+import org.gradle.util.internal.DisconnectableInputStream;
 import org.gradle.internal.io.LineBufferingOutputStream;
 
 import java.io.IOException;
