+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/ContinuousBuildActionExecuter.java
@@ -48,7 +48,7 @@ import org.gradle.internal.time.Clock;
 import org.gradle.launcher.exec.BuildActionExecuter;
 import org.gradle.launcher.exec.BuildActionParameters;
 import org.gradle.launcher.exec.BuildActionResult;
-import org.gradle.util.DisconnectableInputStream;
+import org.gradle.util.internal.DisconnectableInputStream;
 
 import java.util.function.Supplier;
