+++ subprojects/launcher/src/test/groovy/org/gradle/tooling/internal/provider/ContinuousBuildActionExecuterTest.groovy
@@ -37,7 +37,7 @@ import org.gradle.internal.time.Time
 import org.gradle.launcher.exec.BuildActionExecuter
 import org.gradle.launcher.exec.BuildActionParameters
 import org.gradle.test.fixtures.concurrent.ConcurrentSpec
-import org.gradle.util.DisconnectableInputStream
+import org.gradle.util.internal.DisconnectableInputStream
 import org.gradle.util.internal.RedirectStdIn
 import org.junit.Rule
 import spock.lang.Timeout
